complete revamp
This is a complete revamp of the virtual machine. Notable changes:

* Based on Ubuntu 14.04.

* Brightbox packages for Ruby.

* Shell provisioning.

After dealing with Puppet for some time, with modules for apt, concat,
mysql, postgres, ruby, stdlib, and what not, I had enough. This VM is
designed for Rails core contributors and we do not need to learn a DSL
that runs across distros and it is idempotent. It is better for this
particular virtual machine to have one simple bootstrapping shell
script that is procedural and anyone can understand and maintain.

Support for VMware and Parallels has been removed because I cannot
test the new Vagrantfile in these providers. If you really care about
them for Rails **core** development please send a PR to restore them
with the updated configuration. (Rails application development is not
relevant for this VM, please do not send changes motivated by that use
case.)