Rewrite and expand basic sections of the README
I've gone through and expanded the scope of the README a bit to make
starting from scratch easier for new developers. I've clarified how to
read the command blocks in the README, ensured that we're consistent
with using `host $` and `vagrant@rails-dev-box:~$`, and tried to make
the steps easier to follow and understand for someone not already
familiar with how Vagrant works.